Move away from svgo. Doing it on my own. Only resistor kind of working now.

+++ parsing/parsing.go
@@ -29,8 +29,8 @@ type Property struct {
 }
 
 type Action struct {
-	Type  string `'.' @Ident`
-	Units int    `('(' @Int? ')')?`
+	Type  string  `'.' @Ident`
+	Units float64 `('(' @Float? ')')?`
 }
 
 type Value interface{ value() }
